refactor: extract setup command from bin/forge.js to lib/commands/setup.js
Bulk extraction of ~3100 lines (110 functions) from bin/forge.js into
lib/commands/setup.js as a registry-compliant module. bin/forge.js reduced
from 3773 to 584 lines (84.5% reduction).

- All setup functions (interactive, quick, dry-run, agent config, tool
  install, external services) moved to lib/commands/setup.js
- Registry auto-discovers setup.js; bin/forge.js no longer has hard-coded
  setup dispatch
- Added getWorkflowCommands and validateAgents as local functions in setup.js
- Updated test/cli/forge.test.js to check setup.js for structure tests
- Updated test/lazy-dirs.test.js to read setupCoreDocs from new location
- Cleaned unused imports from bin/forge.js (18 requires removed)
- ESLint clean: 0 errors, 0 warnings (--max-warnings 0)
