+++ test/cli/forge.test.js
@@ -1,11 +1,32 @@
-const { describe, test, expect } = require('bun:test');
+const { describe, test, expect } = require('bun:test');
 const fs = require('node:fs');
 const path = require('node:path');
 
 const forgePath = path.join(__dirname, '..', '..', 'bin', 'forge.js');
+const setupPath = path.join(__dirname, '..', '..', 'lib', 'commands', 'setup.js');
 
 describe('bin/forge.js structure', () => {
   const source = fs.readFileSync(forgePath, 'utf8');
+
+  test('should wrap main() in IIFE with try-catch', () => {
+    expect(source.includes('(async () => {')).toBeTruthy();
+    expect(source.includes('await main()')).toBeTruthy();
+  });
+
+  test('should use require.main guard', () => {
+    expect(source.includes('require.main === module')).toBeTruthy();
+  });
+
+  test('should dispatch setup via registry', () => {
+    // Setup is now auto-discovered via lib/commands/setup.js
+    // bin/forge.js should NOT have the inline setup dispatch anymore
+    expect(source.includes("registry.commands.has(command)")).toBeTruthy();
+    expect(source.includes("require('../lib/commands/setup')")).toBeTruthy();
+  });
+});
+
+describe('lib/commands/setup.js structure', () => {
+  const source = fs.readFileSync(setupPath, 'utf8');
 
   test('should export helper functions for cognitive complexity', () => {
     // Verify extracted helpers exist as standalone functions
@@ -40,19 +61,8 @@ describe('bin/forge.js structure', () => {
     }
   });
 
-  test('should wrap main() in IIFE with try-catch', () => {
-    expect(source.includes('(async () => {')).toBeTruthy();
-    expect(source.includes('await main()')).toBeTruthy();
-  });
-
-  test('should use require.main guard', () => {
-    expect(source.includes('require.main === module')).toBeTruthy();
-  });
-
   test('should have Phase 7A helper functions for complexity reduction', () => {
     // Helpers extracted to reduce cognitive complexity in Phase 7A
-    // Note: validateCommonSecurity was extracted to lib/validation-utils.js in CLI maturity epic
-    // Note: detectFromLockFile/detectFromCommand extracted to lib/detection-utils.js
     const phase7aHelpers = [
       'installViaBunx',
       'getSkillsInstallArgs',
@@ -110,10 +120,5 @@ describe('bin/forge.js structure', () => {
     // Verify installBeadsWithMethod and installSkillsWithMethod use the shared helper
     expect(source.includes("installViaBunx('@beads/bd'")).toBeTruthy();
     expect(source.includes("installViaBunx('@forge/skills'")).toBeTruthy();
-  });
-
-  test('should delegate detectPackageManager to detection-utils', () => {
-    // Verify detectPackageManager delegates to lib/detection-utils
-    expect(source.includes("detectionUtils.detectPackageManager(errors, projectRoot)")).toBeTruthy();
   });
 });
